cpp-doxygen refactoring and remove bazel
